+++ js/checkbox/shared/Checkbox.svelte
@@ -5,6 +5,7 @@
 	export let value = false;
 	export let label = "Checkbox";
 	export let interactive: boolean;
+	export let indeterminate = false;
 
 	const dispatch = createEventDispatcher<{
 		change: boolean;
@@ -44,6 +45,7 @@
 <label class:disabled>
 	<input
 		bind:checked={value}
+		bind:indeterminate
 		on:keydown={handle_enter}
 		on:input={handle_input}
 		{disabled}
@@ -102,6 +104,34 @@
 	input:focus {
 		border-color: var(--checkbox-border-color-focus);
 		background-color: var(--checkbox-background-color-focus);
+	}
+
+	input:indeterminate {
+		background-image: none;
+		background-color: var(--checkbox-background-color-selected);
+		border-color: var(--checkbox-border-color-focus);
+		position: relative;
+	}
+
+	input:indeterminate::after {
+		content: "";
+		position: absolute;
+		top: 50%;
+		left: 50%;
+		transform: translate(-50%, -50%);
+		width: 8px;
+		height: 2px;
+		background-color: white;
+	}
+
+	input:indeterminate:hover {
+		background-color: var(--checkbox-background-color-selected);
+		border-color: var(--checkbox-border-color-hover);
+	}
+
+	input:indeterminate:focus {
+		background-color: var(--checkbox-background-color-selected);
+		border-color: var(--checkbox-border-color-focus);
 	}
 
 	input[disabled],