+++ js/core/src/navbar_store.ts
@@ -1,0 +1,9 @@
+import { writable } from "svelte/store";
+
+export interface NavbarConfig {
+	visible?: boolean;
+	main_page_name?: string | false;
+	value?: [string, string][] | null;
+}
+
+export const navbar_config = writable<NavbarConfig | null>(null);