+++ js/spa/src/Index.svelte
@@ -1,6 +1,6 @@
 <script context="module" lang="ts">
 	import { writable } from "svelte/store";
-	import { mount_css as default_mount_css, prefix_css } from "@gradio/core";
+	import { mount_css, prefix_css } from "@gradio/core";
 
 	import type { Client as ClientType } from "@gradio/client";
 
@@ -34,7 +34,7 @@
 		username: string | null;
 		api_prefix?: string;
 		max_file_size?: number;
-		pages: [string, string][];
+		pages: [string, string, boolean][];
 		current_page: string;
 		deep_link_state?: "valid" | "invalid" | "none";
 		page: Record<
@@ -86,8 +86,6 @@
 	import { StatusTracker } from "@gradio/statustracker";
 	import { _ } from "svelte-i18n";
 	import { setupi18n } from "@gradio/core";
-	import type { WorkerProxy } from "@gradio/wasm";
-	import { setWorkerProxyContext } from "@gradio/wasm/svelte";
 	import { init } from "@huggingface/space-header";
 
 	let i18n_ready = false;
@@ -108,22 +106,12 @@
 	export let info: boolean;
 	export let eager: boolean;
 	let stream: EventSource;
-	let pages: [string, string][] = [];
+	let pages: [string, string, boolean][] = [];
 	let current_page: string;
 	let root: string;
 
 	// These utilities are exported to be injectable for the Wasm version.
-	export let mount_css: typeof default_mount_css = default_mount_css;
 	export let Client: typeof ClientType;
-	export let worker_proxy: WorkerProxy | undefined = undefined;
-	if (worker_proxy) {
-		setWorkerProxyContext(worker_proxy);
-
-		worker_proxy.addEventListener("progress-update", (event) => {
-			loading_text = (event as CustomEvent).detail + "...";
-		});
-	}
-	let is_lite = worker_proxy !== undefined;
 
 	export let space: string | null;
 	export let src: string | null;
@@ -532,7 +520,7 @@
 	{pages}
 	{current_page}
 	{root}
-	{is_lite}
+	components={config?.components || []}
 	bind:wrapper
 >
 	{#if i18n_ready}