+++ js/column/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Dependency updates
 
+- @gradio/statustracker@0.11.0
+
+## 0.2.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.18
+
+## 0.2.1
+
+### Dependency updates
+
 - @gradio/statustracker@0.10.17
 
 ## 0.2.1