+++ js/sidebar/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @gradio/sidebar
 
+## 0.1.21
+
+### Dependency updates
+
+- @gradio/atoms@0.17.0
+- @gradio/statustracker@0.11.0
+- @gradio/column@0.2.1
+
+## 0.1.20
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.18
+
 ## 0.1.20
 
 ### Dependency updates